requirements.txt
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
firebase-admin==6.1.0
Flask==2.2.3
Flask-WTF==1.1.1
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.2
msgpack==1.0.5
oauth2client==4.1.3
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-decouple==3.8
python-dotenv==1.0.0
python-jwt==2.0.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==2.3.3
